+++ test/runtests.jl
@@ -14,3 +14,4 @@ end
 
 include("test-get_leaf_nodes.jl")
 include("test-tree-structure.jl")
+include("test-weights.jl")